// src/app/core/models/news.model.ts
export class News {
    sourceID?: any;
    sourceCategoryName?: string;
    id?: number;
    showOnHomepage?: boolean;
    title?: string;
    likeNew?: boolean;
    displayShare?: boolean;
    description?: string;
    urlToImage?: string;
    publishedAt?: any;
    content?: string;
}

export class NewsCategory{
    id?: number;
    name?: string;
}